requirements.txt
fastapi
uvicorn
python-slugify
langchain
langgraph
pydantic
python-multipart